+++ source/menu/menu_config_screen.cpp
@@ -23,8 +23,8 @@ void CMenu::_hideConfigScreen(bool instant)
 	m_btnMgr.hide(m_configScreenLblTVYVal, instant);
 	m_btnMgr.hide(m_configScreenBtnTVYM, instant);
 	m_btnMgr.hide(m_configScreenBtnTVYP, instant);
-	for (u32 i = 0; i < ARRAY_SIZE(m_configScreenLblUser); ++i)
-		if (m_configScreenLblUser[i] != -1u)
+	for(u8 i = 0; i < ARRAY_SIZE(m_configScreenLblUser); ++i)
+		if(m_configScreenLblUser[i] != (u16)-1)
 			m_btnMgr.hide(m_configScreenLblUser[i], instant);
 }
 
@@ -48,8 +48,8 @@ void CMenu::_showConfigScreen(void)
 	m_btnMgr.show(m_configScreenLblTVYVal);
 	m_btnMgr.show(m_configScreenBtnTVYM);
 	m_btnMgr.show(m_configScreenBtnTVYP);
-	for (u32 i = 0; i < ARRAY_SIZE(m_configScreenLblUser); ++i)
-		if (m_configScreenLblUser[i] != -1u)
+	for(u8 i = 0; i < ARRAY_SIZE(m_configScreenLblUser); ++i)
+		if(m_configScreenLblUser[i] != (u16)-1)
 			m_btnMgr.show(m_configScreenLblUser[i]);
 
 	m_btnMgr.setText(m_configScreenLblTVWidthVal, wfmt(L"%i", 640 * 640 / max(1, m_cfg.getInt("GENERAL", "tv_width", 640))));
